Significantly optimize generated C code by preprocessing ourselves
The code is now *significantly* easier for the C compiler to process. The code generated with the old method was pretty fragile and sometimes include depth was deep enough that it was causing the compiler on my local machine (macOS Clang) to segfault (a truly remarkable feat, in my opinion).

The new approach implemented here is to process the C AST trees ourselves and include declarations in the source code directly rather than including a header. Now we only include the most minimal header content possible (runtime and module initialization declarations).