Update Talta; and other important changes...
[runtime] Always free Alta exception values (even non-native values)
I guess I had assumed the destructor would take care of this for us, but that's not the case, because the objects aren't marked as being persistent

[stdlib/string] FREE THE CODEPOINT DATA!!!
This is another one of those things where I'm like "How did I overlook this originally?!".

[stdlib/vector] Delete elements before freeing the underlying array
Again, how did I overlook this? Elements need to be deleted so they can free any resources they might be using *before* we free the array and lose access to them forever.