[stdlib/math] Remove unfinished `Integer` code
I do have plans to eventually implement a "big integer" class for Alta (completely written in Alta), but that'll have to wait until I implement async I/O for files (it's already there for networking). But this is definitely all going into the v0.7.0 release. (Along with a pure-Alta crypto package and *possibly* an HTTPS package).