[internal] Lay the groundwork for sleeping coroutines
Sleeping coroutines are a way to save CPU time and resources because coroutines that need to perform asynch tasks using the platform's native handles can tell the scheduler that they have nothing to do until their handle is ready. The scheduler can then take it off the workqueue, giving other coroutines more chances to run. If *all* coroutines are sleeping (i.e. there are none in the workqueue), the scheduler can save *many* more resources by putting the entire process to sleep until one of the handles is ready (with the user's consent, of course; which will probably be given by waiting for coroutines with a new method like `waitForCoroutines` and a parameter indicating whether to sleep or not).

Note that the part about *platform handles* is important. The only way to make sure we can handle everybody's stuff is if we use *the same interface* for the handles. This is because we can't, for example, wait for mutexes at the same time that we wait for network sockets. `poll` doesn't work with mutexes, and mutexes can't wait for sockets. Maybe I'll find a workaround for this in the future, but for now, only native platform handles can be slept on (because we will use `poll` on Unix-like platforms and `WaitForMultipleObjects` on NT).