[stdlib/io] Improve `File` a bit
It now shares the file handle between copies of the object, which should reduce the toll on the system severely (especially since Alta lacks copy optimization right now and copies everything it sees).

The C file mode has also been abstracted away a bit and simplified. Now it's only a matter of whether you want to read or write and whether you want to overwrite the contents of the file or not.