[stdlib/http] Embrace coroutines (a little bit)
`Server` now schedules coroutines internally to handle client connections.

The HTTP client interface has also been greatly simplified, using an asynchronous function to create a request and wait for it to connect, and then you can use `waitForResponse` to do exactly that: wait for a response asynchronously. It's *significantly* easier than the way we we're previously doing it.