[cli] Only link final targets with the runtime...
...also: double-link the runtime in those final target

This is necessary mainly for Linux, because GNU ld is very lazy and only looks for symbols in preceding libraries, and then immediately forgets them. So we need to specify the runtime twice: once *before* the core target because the `_internal` module depends on it, and once *after* the core target because the runtime depeds on the `_internal` module.