[stdlib] Start removing dependency on C transpiler
I'm planning on rewriting Talta to use LLVM to generate machine code
instead of transpiling to C (because keeping track of includes and
declaration order is a nightmare). In order to do so, we first need to
start providing our own declarations for things we previously expected
to be able to include from a C header. This is perfectly fine to do
since such definitions/declarations are part of the C ABI for each
platform and, as such, they won't change much (if at all)
 it's just very tedious :/